move ArchiveDay interface to new Interfaces namespace + create ArchiveService interface
